pinky0: moved closer to homerow key
The hight is raised by 1.4mm and the switch is moved toward the
center by 2mm. With the steep surface slope (about 45 degrees) it
lowers the key as well. So, by raising the segment hight and lowering
the switch (while moving toward the homerow key) we should keep the
switch at the same height while bringing it closer to the homerow pinky
key.