pinky2_length: increased by 3mm
The intent is to move the switch away from the center as it touches the
pinky1 switch. As there is no room left to move the switch_center safely (the
switch would be pressed against the side wall), we have to
increase the segment length by 3mm.
This should move the switch_center by 1.5mm away from the
center.
However, this will also decrease the switch surface angle.